Fixed the "byte alignment" issue with DataStruct<int> and DataStruct<size_t> in GridPatch.h:
DataArray1D<size_t> m_iActiveStatePatchIx;
DataArray1D<size_t> m_iGeometricPatchIx;

This also needed a fix in OutputManagerComposite.cpp on lines 260 and
277 to type cast:

MPI_Isend(
				const_cast<unsigned char *>(pGeometricData),
				nGeometricDataByteSize,
				MPI_BYTE,
				0,
				ExchangeDataType_Geometric,
				MPI_COMM_WORLD,
				&(vecSendReqGeo[i]));
				
MPI_Isend(
				const_cast<unsigned char *>(pActiveStateData),
				nActiveStateDataByteSize,
				MPI_BYTE,
				0,
				ExchangeDataType_ActiveState,
				MPI_COMM_WORLD,
				&(vecSendReqAcS[i]));
				
Updated the BC's, Cartesian test cases, and ARK2 and ARK3. JEG

+++ src/atm/GridPatch.h
@@ -1081,7 +1081,7 @@ protected:
 	///	<summary>
 	///		Geometric patch index.
 	///	</summary>
-	DataArray1D<int> m_iGeometricPatchIx;
+	DataArray1D<size_t> m_iGeometricPatchIx;
 
 	///	<summary>
 	///		Array of nodal values in the alpha direction (Geometric).
@@ -1254,7 +1254,7 @@ public:
 	///	<summary>
 	///		Active State patch index.
 	///	</summary>
-	DataArray1D<int> m_iActiveStatePatchIx;
+	DataArray1D<size_t> m_iActiveStatePatchIx;
 
 	///	<summary>
 	///		Grid data for the reference state on model levels (State).
